+++ Changelog.md
@@ -7,6 +7,7 @@
 ## Features
 * Show spinner on initial stream load [#6384](https://github.com/diaspora/diaspora/pull/6384)
 * Add new moderator role. Moderators can view and act on reported posts [#6351](https://github.com/diaspora/diaspora/pull/6351)
+* Only post to the primary tumblr blog [#6386](https://github.com/diaspora/diaspora/pull/6386)
 
 # 0.5.3.0